requirements.txt
Fastapi
uvicorn
python-multipart
pyyaml
cryptography
python-dotenv
boto3
mangum